# api.yaml
swagger: '2.0'
info:
  title: API микросервиса управления договорами страхования
  version: 1.0.0
paths:
...
  /insurance-service/v2/changereport/generate:
    get:
      tags:
        - ChangeReport
      description: Создать отчёт по изменениям от определённой даты в папке с отчётами
      operationId: generateChangeReport
      produces:
        - application/json
      parameters:
        - in: query
          name: createDate
          type: string
          format: date
          required: true
          description: Дата (С) создания изменения
        - in: query
          name: endDate
          type: string
          format: date
          description: Дата (ПО) создания изменения
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/uuidRs'
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
...
  /insurance-service/v1/dict/requiredDocuments:
    get:
      tags:
        - RequiredDocuments
      description: Получить список всех записей справочника обязательных документов.
      operationId: RequiredDocuments
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            description: Постраничный вывод данных справочника
            $ref: '#/definitions/requiredDocumentData'
        403:
          description: Forbidden
        500:
          description: Internal server error
    post:
      tags:
        - RequiredDocuments
      description: Создать запись в справочнике обязательных документов.
      operationId: createRequiredDocument
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: createRequiredDocumentRq
          description: Запрос создания записи в справочнике обязательных документов.
          required: true
          schema:
            $ref: '#/definitions/requiredDocumentData'
      responses:
        200:
          description: OK
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        403:
          description: Forbidden
        500:
          description: Internal server error
  /insurance-service/v1/dict/requiredDocuments/{id}:
    get:
      tags:
        - RequiredDocuments
      description: Получить запись из справочника обязательных документов.
      operationId: getRequiredDocument
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи в справочнике обязательных документов.
      responses:
        200:
          description: OK
          schema:
            description: Данные записи обязательного документа.
            $ref: '#/definitions/requiredDocumentData'
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
    put:
      tags:
        - RequiredDocuments
      description: Обновить запись в справочнике обязательных документов.
      operationId: putRequiredDocument
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи в справочнике обязательных документов.
        - in: body
          name: updateRequiredDocumentsRq
          description: Запрос обновления записи обязательного документа.
          required: true
          schema:
            $ref: '#/definitions/requiredDocumentData'
      responses:
        200:
          description: OK
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
    delete:
      tags:
        - RequiredDocuments
      description: Удалить запись в справочнике обязательных документов.
      operationId: deleteRequiredDocuments
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи обязательного документа.

      responses:
        200:
          description: OK
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        403:
          description: Forbidden
        404:
          description: Not Found
        500:
          description: Internal server error
  /insurance-service/v1/dict/strategies:
    get:
      tags:
        - Strategies
      description: Получить список всех записей справочника стратегий.
      operationId: getStrategyList
      x-spring-pageable: true
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            description: Постраничный вывод данных справочника
            $ref: '#/definitions/strategyData'
        403:
          description: Forbidden
        500:
          description: Internal server error
    put:
      tags:
        - Strategies
      description: Получить список записей справочника стратегий по коэффициенту участия.
      operationId: getStrategyListByRate
      produces:
        - application/json
      parameters:
        - in: query
          type: number
          name: coefficient
          description: Коэффициент участия
          required: true
      responses:
        200:
          description: OK
          schema:
            description: Вывод данных справочника
            type: array
            items:
              $ref: '#/definitions/strategyData'
        403:
          description: Forbidden
        500:
          description: Internal server error
  /insurance-service/v1/dict/strategy:
    post:
      tags:
        - Strategies
      description: Создать запись справочника стратегий
      operationId: createStrategy
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: createStrategyRq
          description: Запрос создания записи в справочнике стратегий.
          required: true
          schema:
            $ref: '#/definitions/strategyData'
      responses:
        200:
          description: OK
          schema:
            description: Стратегия программ страхования из справочника
            $ref: '#/definitions/strategyData'
        403:
          description: Forbidden
        500:
          description: Internal server error
  /insurance-service/v1/dict/strategy/{id}:
    get:
      tags:
        - Strategies
      description: Получить запись справочника стратегий по идентификатору
      operationId: getStrategyById
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи в справочнике стратегий.
      responses:
        200:
          description: OK
          schema:
            description: Стратегия программ страхования из справочника
            $ref: '#/definitions/strategyData'
        403:
          description: Forbidden
        500:
          description: Internal server error
    put:
      tags:
        - Strategies
      description: Обновить запись справочника стратегий по идентификатору
      operationId: updateStrategyById
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи в справочнике стратегий.
        - in: body
          name: updateStrategyRq
          description: Запрос обновления записи в справочнике стратегий.
          required: true
          schema:
            $ref: '#/definitions/strategyData'
      responses:
        200:
          description: OK
          schema:
            description: Стратегия программ страхования из справочника
            $ref: '#/definitions/strategyData'
        403:
          description: Forbidden
        500:
          description: Internal server error
    delete:
      tags:
        - Strategies
      description: Удалить запись справочника стратегий по идентификатору
      operationId: deleteStrategy
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: id
          description: Идентификатор записи в справочнике стратегий.
      responses:
        200:
          description: OK
        403:
          description: Forbidden
        500:
          description: Internal server error
...
definitions:
...
  errorModel:
    type: object
    description: Описание ошибки
    required:
      - errors
    properties:
      errors:
        type: array
        items:
          type: string
...
  requiredDocumentData:
    type: object
    description: Данные обязательного документа
    required:
      - type
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи
      type:
        type: string
        description: Тип документа
      activeFlag:
        type: boolean
        description: Признак действующего документа
...
  strategyData:
    type: object
    description: Данные записи справочника стратегий
    required:
      - program
      - name
      - rate
      - policyCode
      - startDate
      - endDate
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи
      name:
        type: string
        maxLength: 50
        description: Вид программы страхования
      description:
        type: string
        maxLength: 500
        description: Описание
      policyCode:
        type: integer
        description: Кодировка полиса стратегия. Числовое поле формата 99, где 9 – цифра от нуля до 9. Данное поле участвует в формировании номера договора
      startDate:
        type: string
        format: date
        description: Дата начала действия стратегии. По умолчанию, текущая дата.
      endDate:
        type: string
        format: date
        description: Дата окончания действия стратегии. По умолчанию, 31.12.2999
      strategyProperties:
        type: array
        items:
          $ref: '#/definitions/strategyProperty'
      type:
        description: Тип ИСЖ стратегии
        $ref: '#/definitions/strategyType'
  strategyType:
    type: string
    description: Виды типов стратегии
    enum:
      - CLASSIC
      - COUPON
      - LOCOMOTIVE
  strategyProperty:
    type: object
    description: Наборы данных стратегии - Дата экспирации, дата НЗБА и тикер -необязательные. коэффициент участия - обязательный.
    required:
      - coefficient
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор записи
      coefficient:
        type: number
        description: Коэффициент участия или размер инвестиционного купона. Значение задается в процентах
      ticker:
        type: string
        maxLength: 20
        description: Тикер. Необязательно для заполнения. Отображается только в печатных формах
      expirationDate:
        type: string
        format: date
        description: Дата экспирации
      nzbaDate:
        type: string
        format: date
        description: Дата НЗБА (начальное значение базового актива)
...
  uuidRs:
    type: object
    description: Ответ, содержащий uuid объекта или процесса
    properties:
      uuid:
        type: string
        description: uuid сущности или процесса
...